some abstraction to util.js and repl.js
