+++ package.json
@@ -46,5 +46,9 @@
   "pkg": {
     "scripts": "build/contracts/*.json",
     "assets": "assets/*.json"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-preset-env": "^1.6.1"
   }
 }
